CSock::GetPeerFingerprint returns long not int
../Csocket.cc:2168:36: error: conversion to ‘int’ from ‘long int’ may
alter its value [-Werror=conversion]

+++ Csocket.cc
@@ -2144,7 +2144,7 @@ CS_STRING Csock::GetPeerPubKey()
 	}
 	return( sKey );
 }
-int Csock::GetPeerFingerprint( CS_STRING & sFP )
+long Csock::GetPeerFingerprint( CS_STRING & sFP )
 {
 	sFP.clear();
